Adds Playwright: e2e and external url tests (#1533)
