+++ .github/workflows/node.js.yml
@@ -26,3 +26,4 @@ jobs:
     - run: npm install
     - run: npm run build
     - run: npm test
+    - run: npm run playwright
