+++ frontend/package.json
@@ -55,6 +55,9 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
+    "e2e": "npx playwright install chromium && npx playwright test --project E2E",
+    "url": "npx playwright install chromium && npx playwright test --project URL",
+    "playwright": "npx playwright install chromium && npx playwright test",
     "eject": "react-scripts eject",
     "storybook": "start-storybook -p 6006 -s public",
     "build-storybook": "build-storybook -s public",
@@ -101,6 +104,7 @@
     ]
   },
   "devDependencies": {
+    "@playwright/test": "^1.20.2",
     "@storybook/addon-actions": "^6.2.8",
     "@storybook/addon-essentials": "^6.2.8",
     "@storybook/addon-links": "^6.2.8",
