+++ frontend/src/pages/AboutUs/OurTeamData.ts
@@ -1,0 +1,250 @@
+export const LEADERSHIP_TEAM = [
+  {
+    name: "Daniel Dawes, JD",
+    role: "Principal Investigator",
+    imageUrl: "/img/team/Daniel-E 1.png",
+  },
+  {
+    name: "Nelson Dunlap, JD",
+    role: "Co-Principal Investigator",
+    imageUrl: "/img/team/DunlapNelson.png",
+  },
+  {
+    name: "Ebony Respress, MPH",
+    role: "Project Director",
+    imageUrl: "/img/team/RespressEbony.png",
+  },
+  {
+    name: "Allyson Belton, MPH",
+    role: "Coalition/Engagement",
+    imageUrl: "/img/team/Belton_Allyson.png",
+  },
+  {
+    name: "Mahia Valle, MBA",
+    role: "Communications",
+    imageUrl: "/img/team/ValleMahia.png",
+  },
+  {
+    name: "Shaneeta M. Johnson MD, MBA, FACS, FASMBS, ABOM",
+    role: "Senior Advisor",
+    imageUrl: "/img/team/ShaneetaJohnson.png",
+  },
+  {
+    name: "JC Gonzalez, MBA, PMP",
+    role: "Product Manager",
+    imageUrl: "/img/team/GonzalezJC.png",
+  },
+  {
+    name: "Josh Zarrabi",
+    role: "Senior Software Engineer",
+    imageUrl: "/img/team/ZarrabiJosh.png",
+  },
+  {
+    name: "Ben Hammond",
+    role: "Software Engineer",
+    imageUrl: "/img/team/HammondBen.jpg",
+  },
+  {
+    name: "Maisha Standifer, PhD, MPH",
+    role: "Health Policy Analyst",
+    imageUrl: "/img/team/maisha-standifer.jpg",
+  },
+  {
+    name: "Jammie Hopkins, DrPH, MS",
+    role: "Coalition/Engagement",
+    imageUrl: "/img/team/jammie-hopkins.jpg",
+  },
+  {
+    name: "Megan Douglas, JD",
+    role: "Health Policy Analyst",
+    imageUrl: "/img/team/DouglasMegan.png",
+  },
+];
+
+export const GOOGLE_FELLOWS = [
+  {
+    name: "Chelsea Seabron",
+    role: "Google.org Manager",
+  },
+  {
+    name: "Rasmi Elasmar",
+    role: "Google.org Manager",
+  },
+  {
+    name: "Jen Carter",
+    role: "Google.org Manager",
+  },
+  {
+    name: "Colin Jackson",
+    role: "Product Manager",
+  },
+  {
+    name: "Larry Adams",
+    role: "Product Manager",
+  },
+  {
+    name: "Dawn Shaikh",
+    role: "UX Researcher",
+  },
+  {
+    name: "Ken Moore",
+    role: "UX Designer",
+    link: "https://kenstatus.com",
+    text: "In memoriam,",
+  },
+  {
+    name: "Chi Pham",
+    role: "UX Designer",
+  },
+  {
+    name: "Katrina Sostek",
+    role: "Engineer",
+  },
+  {
+    name: "Vansh Kumar",
+    role: "Engineer",
+  },
+  {
+    name: "Lorenzo Boyice",
+    role: "Engineer",
+  },
+  {
+    name: "Jennie Brown",
+    role: "Engineer",
+  },
+  {
+    name: "Krista Katzenmeyer",
+    role: "Engineer",
+  },
+  {
+    name: "Aaron Nelson",
+    role: "Engineer",
+  },
+  {
+    name: "Maya Spivak",
+    role: "Engineer",
+  },
+  {
+    name: "Brian Suk",
+    role: "Engineer",
+  },
+  {
+    name: "Elizabeth Fregoso",
+    role: "Engineer",
+  },
+  {
+    name: "Seth Van Grinsven",
+    role: "Engineer",
+  },
+  {
+    name: "Alexis Artis",
+    role: "Marketing",
+  },
+  {
+    name: "Tarah Moore",
+    role: "Program Manager",
+  },
+  {
+    name: "Akoua Smith",
+    role: "Brand Designer",
+  },
+  {
+    name: "Sohee Kim",
+    role: "Brand Designer",
+  },
+  {
+    name: "Emily Rothschild",
+    role: "Content Strategist",
+  },
+];
+
+export const HE_TASKFORCE = [
+  "Douglas M. Brooks",
+  "Duanne Pearson",
+  "David Introcaso, PhD",
+  "Zinzi Bailey, ScD, MSPH",
+  "Abigail Echo-Hawk, MA",
+  "Luis Belén",
+  "Joia Crear-Perry, MD, FACOG",
+  "Jenné Johns, MPH",
+  "Erica Taylor, PhD, MPH, MA",
+  "Brian Smedley, PhD",
+  "Leandris Liburd, PhD, MPH",
+  "Rita Carreón",
+  "Dr. Raynald Samoa, MD",
+  "Janisse Rosario Schoepp, MPH, PhD",
+  "Elena Ong, PHN, MS",
+  "J. Nadine Gracia, MD, MSCE",
+  "Todd Moore, MPS",
+  "Francys Crevier, JD",
+  "Elena Rios, MD, MSPH, FACP",
+  "Coleen A. Boyle, PhD, MS hyg",
+  "Pierluigi Mancini, PhD",
+  "Judy Monroe, MD",
+  "Kathy Ko Chin, MS",
+  "Marcos Pesquera, RPh, MPH",
+  "Tawara D. Goode, MA",
+  "Brian C. Castrucci, DrPH, MA",
+  "The Honorable Donna M. Christensen, MD",
+  "Dawn M. Hunter, JD, MPH",
+  "Eduardo J. Gomez, PhD, MA",
+  "Marjorie A. Innocent, PhD",
+  "Geoffrey M. Roche, MPA",
+  "Lori Tremmel Freeman, MBA",
+  "Roberta Carlin, MS, JD",
+  "Scott Bryant Comstock, MS",
+  "Ruqaiijah A. Yearby, JD, MPH",
+  "Magda G Peck, ScD",
+  "Aletha Maybank, MD, MPH",
+  "RADM Michael Toedt, MD, FAAFP",
+  "Andrea Collier",
+  "Kimberlydawn Wisdom, MD, MS",
+  "Philip Alberti PhD",
+  "Robin W. Coleman",
+  "Corinne M. Graffunder, DrPH",
+  "T'Pring Westbrook, PhD",
+  "Fernando De Maio, PhD",
+  "Mary Ann Cooney, MPH, MSN",
+  "Ngozi Afulezi, MHA",
+  "Alyssa Cobb",
+  "Lauren Smith, MD, MPH",
+  "Helene Clayton-Jeter, OD",
+  "Richard Calvin Chang, JD",
+  "Dr. Joseph Sakran, MD, MPA, MPH",
+  "Darrell LaRoche",
+  "Shavon Arline-Bradley, MPH, MDiv",
+  "Melicia Whitt-Glover, PhD, FACSM",
+];
+
+export const PARTNERS = [
+  {
+    imageUrl: "/img/partners/PartnerSatcher.png",
+    alt: "Morehouse School of Medicine Satcher Health Leadership Institute",
+    url: "https://satcherinstitute.org",
+  },
+  {
+    imageUrl: "/img/partners/PartnerGilead.png",
+    alt: "Gilead Sciences, Inc.",
+    url: "https://www.gilead.com/",
+  },
+  {
+    imageUrl: "/img/partners/PartnerCdc.png",
+    alt: "United States Center for Disease Control and Prevention",
+    url: "https://www.cdc.gov/",
+  },
+  {
+    imageUrl: "/img/partners/PartnerGoogle.png",
+    alt: "Google.org",
+    url: "https://google.org",
+  },
+  {
+    imageUrl: "/img/partners/PartnerAARP.png",
+    alt: "American Association of Retired Persons",
+    url: "https://www.aarp.org",
+  },
+  {
+    imageUrl: "/img/partners/PartnerAECF.png",
+    alt: "Annie E. Casey Foundation",
+    url: "https://www.aecf.org/",
+  },
+];
