+++ frontend/src/reports/OneVariableReport.tsx
@@ -129,7 +129,7 @@ export function OneVariableReport(props: OneVariableReportProps) {
           </Grid>
 
           {/* 100k MAP CARD */}
-          <Grid item xs={12} md={SINGLE_COLUMN_WIDTH}>
+          <Grid item xs={12} md={SINGLE_COLUMN_WIDTH} id="map">
             <MapCard
               variableConfig={variableConfig}
               fips={props.fips}
@@ -143,7 +143,7 @@ export function OneVariableReport(props: OneVariableReportProps) {
           </Grid>
 
           {/* 100K BAR CHART CARD */}
-          <Grid item xs={12} sm={12} md={SINGLE_COLUMN_WIDTH}>
+          <Grid item xs={12} sm={12} md={SINGLE_COLUMN_WIDTH} id="bar">
             <LazyLoad offset={600} height={750} once>
               {DEMOGRAPHIC_BREAKDOWNS.map((breakdownVar) => (
                 <Fragment key={breakdownVar}>
@@ -161,7 +161,7 @@ export function OneVariableReport(props: OneVariableReportProps) {
           </Grid>
 
           {/* UNKNOWNS MAP CARD */}
-          <Grid item xs={12} sm={12} md={SINGLE_COLUMN_WIDTH}>
+          <Grid item xs={12} sm={12} md={SINGLE_COLUMN_WIDTH} id="unknowns">
             <LazyLoad offset={800} height={750} once>
               {variableConfig.metrics["pct_share"] && (
                 <UnknownsMapCard
@@ -178,7 +178,7 @@ export function OneVariableReport(props: OneVariableReportProps) {
           </Grid>
 
           {/* DISPARITY BAR CHART COMPARE VS POPULATION */}
-          <Grid item xs={12} sm={12} md={SINGLE_COLUMN_WIDTH}>
+          <Grid item xs={12} sm={12} md={SINGLE_COLUMN_WIDTH} id="disparity">
             <LazyLoad offset={800} height={750} once>
               {DEMOGRAPHIC_BREAKDOWNS.map((breakdownVar) => (
                 <Fragment key={breakdownVar}>
@@ -196,7 +196,7 @@ export function OneVariableReport(props: OneVariableReportProps) {
           </Grid>
 
           {/* DATA TABLE CARD */}
-          <Grid item xs={12} md={SINGLE_COLUMN_WIDTH}>
+          <Grid item xs={12} md={SINGLE_COLUMN_WIDTH} id="table">
             <LazyLoad offset={800} height={750} once>
               {DEMOGRAPHIC_BREAKDOWNS.map((breakdownVar) => (
                 <Fragment key={breakdownVar}>
@@ -213,7 +213,7 @@ export function OneVariableReport(props: OneVariableReportProps) {
           </Grid>
 
           {/* AGE ADJUSTED TABLE CARD */}
-          <Grid item xs={12} md={SINGLE_COLUMN_WIDTH}>
+          <Grid item xs={12} md={SINGLE_COLUMN_WIDTH} id="age-adjusted">
             <LazyLoad offset={800} height={800} once>
               <AgeAdjustedTableCard
                 fips={props.fips}
